chore: more component modularization
