+++ src/components/LogoComposite/LogoComposite.scss
@@ -1,0 +1,28 @@
+.logo-composite {
+  margin-top: 6rem;
+  display: flex;
+  align-items: center;
+
+  &__divider {
+    height: 4rem;
+    margin: 0 2rem;
+    border-left-width: 1px;
+    border-style: solid;
+    border-color: var(--element-border-color);
+  }
+
+  &__logo {
+    width: 12rem;
+    max-width: 90%;
+    min-height: 54px;
+    padding-bottom: 1rem;
+  }
+
+  &__subtitle {
+    color: var(--oc-gray-5);
+    font-size: 3rem;
+    font-weight: 100;
+    letter-spacing: 0.275em;
+    line-height: 0;
+  }
+}
