+++ src/styles.scss
@@ -1,33 +1,36 @@
 // Define variable defaults
-@import '~open-color/open-color.scss';
-
-$itemis-dark-blue: hsl(206, 100%, 8%);
-
-$primary: hsl(206, 100%, 36%);
-
-$body-bg: $itemis-dark-blue;
-$body-color: $oc-gray-4;
-
-$input-bg: hsl(206, 100%, 12%);
-$input-color: $oc-gray-4;
-$input-border-color: hsl(206, 100%, 16%);
-
-$secondary: #dbe2f0;
+@import '~open-color/open-color.css';
 
 // Then import Bootstrap and BootstrapVue SCSS files (order is important)
 @import '~bootstrap/scss/bootstrap.scss';
 @import '~bootstrap-vue/src/index.scss';
 
+$itemis-dark-blue: hsl(206, 100%, 8%);
+
 :root {
-  --itemis-brand-dark-bg: $itemis-dark-blue;
-  --itemis-nav-dark: hsl(206, 100%, 10%);
-  --itemis-nav-dark-border: hsl(206, 100%, 16%);
+  --primary-color: hsl(206, 100%, 36%);
+  --secondary-color: #dbe2f0;
+
+  --bg-color: hsl(206, 100%, 8%);
+  --text-color: var(--oc-gray-4);
+
+  --element-bg-color: hsl(206, 100%, 10%);
+  --element-border-color: hsl(206, 100%, 16%);
+
+  --input-text-color: var(--oc-gray-4);
+  --input-bg-color: hsl(206, 100%, 12%);
+  --input-border-color: hsl(206, 100%, 16%);
 }
 
 body {
-  min-height: 100vh;
+  min-height: 100vh !important;
+  background-color: var(--bg-color) !important;
+  color: var(--text-color) !important;
 }
 
 input {
+  border-color: var(--input-border-color) !important;
+  background-color: var(--input-bg-color) !important;
+  color: var(--input-text-color) !important;
   font-size: inherit !important;
 }
